move evaluation to binder and add test cases
