+++ e2e_test/batch/order/negative_offset.slt.part
@@ -15,32 +15,6 @@ SELECT * FROM generate_series(0,10,1) LIMIT -3;
 statement error
 SELECT * FROM generate_series(0,10,1) LIMIT -1%;
 
-query I
-SELECT * FROM generate_series(0,10,1) as t(v) order by v limit 1+2;
-----
-0
-1
-2
-
-query I
-SELECT * FROM generate_series(0,10,1) as t(v) order by v limit 1.5*2;
-----
-0
-1
-2
-
-query I
-SELECT * FROM generate_series(0,10,1) as t(v) order by v limit -3+4;
-----
-0
-
-query I
-SELECT * FROM generate_series(0,10,1) as t(v) order by v limit 1.5*2 OFFSET 2;
-----
-2
-3
-4
-
 statement ok
 CREATE TABLE integers(k int);
